transacciones/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
requests
kafka-python